split screens to different files
