+++ src/gui/common.py
@@ -1,3 +1,4 @@
+"""Some commonly used functions, like helpers"""
 import lvgl as lv
 import qrcode
 import math
