+++ src/wallets/__init__.py
@@ -1,1 +1,2 @@
 from .manager import WalletManager
+from .base import WalletError
