+++ src/wallets/base.py
@@ -1,0 +1,5 @@
+class WalletError(Exception):
+    pass
+
+class Wallet:
+	pass
